requirements.txt
paramiko==4.0.0
PyYAML==6.0.3
requests==2.32.5
kubernetes==34.1.0
Flask==3.1.2
cryptography==46.0.3